separate leagueUserType to its own file
