+++ src/components/leagueUser/league-user-types.ts
@@ -1,0 +1,7 @@
+import { type LeagueMatch, type TeamUser } from "@prisma/client";
+
+export interface LeagueMatchWithProfiles {
+  winnerTeamUser: TeamUser;
+  loserTeamUser: TeamUser;
+  match: LeagueMatch;
+}
